Make database path dynamically changeable
This commit refactors the way the database path is resolved and used by
moro's db-driver. Instead of manually combining the default name to the
home-dir - which is the default by the way - the config file is
consulted first and it's path is used if it's not empty. If it's empty
the default db path is used.

- Read db path from config file if possible, fallback to default path.
- Explicitly setup the db path of the test db.
- Use resolved db file path for removal, instead of using the given one
  via parameter.
- Remove unnecessary `console.log` call.

Additionally adapt tests to function with the current setup:
- Install and use `cross-env` to setup test mode.
- Use exported `dbTestFile` instead of manually creating it.
- Fix removal test to use new signature of `removeDatabase`.